refactor: break up existing code into many files
